ai-service/requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
pydantic==2.*
openai==0.28.*
python-dotenv==1.*
